refactor(collavre): extract SessionProvisioner and TaskClaimService
Decompose Api::V1::AgentsController by moving two cohesive responsibilities
into AiAgent services:
- SessionProvisioner: register-time agent + session-topic provisioning
  (find_or_create_agent/topic, deterministic email, ownership verification).
- TaskClaimService: atomic delegated-task claim + post-save finalize.

The controller keeps a thin claim_delegated_task delegator so the concurrency
test that patches that method keeps its seam. A2A dispatch already lives in the
dedicated AiAgent::A2aDispatcher service, so no new extraction there.
